# openapi.json
{"openapi": "3.1.0", "info": {"title": "Huldra Dashboard", "description": "Monitoring dashboard for Huldra experiments", "version": "0.1.0"}, "paths": {"/api/health": {"get": {"tags": ["api"], "summary": "Health Check", "description": "Health check endpoint.", "operationId": "health_check_api_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthCheck"}}}}}}}, "/api/experiments": {"get": {"tags": ["api"], "summary": "List Experiments", "description": "List all experiments with optional filtering.", "operationId": "list_experiments_api_experiments_get", "parameters": [{"name": "result_status", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "Filter by result status", "title": "Result Status"}, "description": "Filter by result status"}, {"name": "attempt_status", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "Filter by attempt status", "title": "Attempt Status"}, "description": "Filter by attempt status"}, {"name": "namespace", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "Filter by namespace prefix", "title": "Namespace"}, "description": "Filter by namespace prefix"}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 1000, "minimum": 1, "description": "Maximum number of results", "default": 100, "title": "Limit"}, "description": "Maximum number of results"}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "description": "Offset for pagination", "default": 0, "title": "Offset"}, "description": "Offset for pagination"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExperimentList"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/experiments/{namespace}/{hexdigest}": {"get": {"tags": ["api"], "summary": "Get Experiment", "description": "Get detailed information about a specific experiment.", "operationId": "get_experiment_api_experiments__namespace___hexdigest__get", "parameters": [{"name": "namespace", "in": "path", "required": true, "schema": {"type": "string", "title": "Namespace"}}, {"name": "hexdigest", "in": "path", "required": true, "schema": {"type": "string", "title": "Hexdigest"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExperimentDetail"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/stats": {"get": {"tags": ["api"], "summary": "Dashboard Stats", "description": "Get aggregate statistics for the dashboard.", "operationId": "dashboard_stats_api_stats_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DashboardStats"}}}}}}}}, "components": {"schemas": {"DashboardStats": {"properties": {"total": {"type": "integer", "title": "Total"}, "by_result_status": {"items": {"$ref": "#/components/schemas/StatusCount"}, "type": "array", "title": "By Result Status"}, "by_attempt_status": {"items": {"$ref": "#/components/schemas/StatusCount"}, "type": "array", "title": "By Attempt Status"}, "running_count": {"type": "integer", "title": "Running Count"}, "queued_count": {"type": "integer", "title": "Queued Count"}, "failed_count": {"type": "integer", "title": "Failed Count"}, "success_count": {"type": "integer", "title": "Success Count"}}, "type": "object", "required": ["total", "by_result_status", "by_attempt_status", "running_count", "queued_count", "failed_count", "success_count"], "title": "DashboardStats", "description": "Aggregate dashboard statistics."}, "ExperimentAttempt": {"properties": {"id": {"type": "string", "title": "Id"}, "number": {"type": "integer", "title": "Number"}, "backend": {"type": "string", "title": "Backend"}, "status": {"type": "string", "title": "Status"}, "started_at": {"type": "string", "title": "Started At"}, "heartbeat_at": {"type": "string", "title": "Heartbeat At"}, "lease_expires_at": {"type": "string", "title": "Lease Expires At"}, "owner": {"$ref": "#/components/schemas/ExperimentOwner"}, "ended_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Ended At"}, "reason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reason"}}, "type": "object", "required": ["id", "number", "backend", "status", "started_at", "heartbeat_at", "lease_expires_at", "owner"], "title": "ExperimentAttempt", "description": "Attempt information for an experiment."}, "ExperimentDetail": {"properties": {"namespace": {"type": "string", "title": "Namespace"}, "hexdigest": {"type": "string", "title": "Hexdigest"}, "class_name": {"type": "string", "title": "Class Name"}, "result_status": {"type": "string", "title": "Result Status"}, "attempt_status": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Attempt Status"}, "attempt_number": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Attempt Number"}, "updated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Updated At"}, "started_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Started At"}, "directory": {"type": "string", "title": "Directory"}, "state": {"additionalProperties": true, "type": "object", "title": "State"}, "metadata": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Metadata"}, "attempt": {"anyOf": [{"$ref": "#/components/schemas/ExperimentAttempt"}, {"type": "null"}]}}, "type": "object", "required": ["namespace", "hexdigest", "class_name", "result_status", "directory", "state"], "title": "ExperimentDetail", "description": "Detailed experiment information."}, "ExperimentList": {"properties": {"experiments": {"items": {"$ref": "#/components/schemas/ExperimentSummary"}, "type": "array", "title": "Experiments"}, "total": {"type": "integer", "title": "Total"}}, "type": "object", "required": ["experiments", "total"], "title": "ExperimentList", "description": "List of experiments with total count."}, "ExperimentOwner": {"properties": {"pid": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Pid"}, "host": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Host"}, "hostname": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Hostname"}, "user": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "User"}, "command": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Command"}, "timestamp": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Timestamp"}}, "type": "object", "title": "ExperimentOwner", "description": "Owner information for an experiment attempt."}, "ExperimentSummary": {"properties": {"namespace": {"type": "string", "title": "Namespace"}, "hexdigest": {"type": "string", "title": "Hexdigest"}, "class_name": {"type": "string", "title": "Class Name"}, "result_status": {"type": "string", "title": "Result Status"}, "attempt_status": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Attempt Status"}, "attempt_number": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Attempt Number"}, "updated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Updated At"}, "started_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Started At"}}, "type": "object", "required": ["namespace", "hexdigest", "class_name", "result_status"], "title": "ExperimentSummary", "description": "Summary of an experiment for list views."}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HealthCheck": {"properties": {"status": {"type": "string", "title": "Status"}, "version": {"type": "string", "title": "Version"}}, "type": "object", "required": ["status", "version"], "title": "HealthCheck", "description": "Health check response."}, "StatusCount": {"properties": {"status": {"type": "string", "title": "Status"}, "count": {"type": "integer", "title": "Count"}}, "type": "object", "required": ["status", "count"], "title": "StatusCount", "description": "Count of experiments by status."}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}